Miscellaneous cleanup,put read/writesync in the control api, warnings, valgrind issue, debug prints Makefile: build transport lib before msg drivers to fix dependency when opensplice/dds is available Makefile.ocpi.common: remove extra slash adapt/os/ocpios/linux/src/OcpiOsDebug.cxx: namespace fixup, add mutex in log printing components/specs/bias_spec.xml: remove unnecessary xml Properties core/container/ctests/bin/src/run_tests.csh: adjust for new log printing to stderr core/container/hdl_container/impl/src/HdlContainer.cxx: warnings core/container/interfaces/api/include/OcpiContainerApi.h: add readsync/writesync to control api core/container/interfaces/api/src/OcpiProperty.cxx: ditto core/container/interfaces/api/src/OcpiContainerApplication.cxx: another printf to log print core/container/interfaces/api/src/OcpiMetadataPort.cxx: printf->logprint core/container/rcc_container/src/OcpiRccWorker.cxx: debug print core/dataplane/msg_driver_interface/include/DtMsgDriver.h: make protocol const, avoid copy ---- this fixed the valgrind issue, but more analysis is needed for lifecycle of protocol object core/dataplane/msg_drivers/Makefile.ocpi.imports: link hard against opensplice when it is there core/dataplane/transport/impl/src/OcpiPort.cxx: debug print fix core/local/util/misc/include/OcpiUtilMisc.h: debug prints fixed core/local/util/property/src/OcpiUtilProtocol.cxx: fix, but also comment out copy constructors core/sca/srpl/srpl/src/Cp289GenericProxy.cxx: use new readsync/writesync api install_opensplice.sh: csh/sh typo tools/cdk/examples/hello_xml/appHello.cxx: give thread time to work tools/cdk/include/hdl/hdl-pre.mk: force remove for sim target files tools/cdk/include/hdl/verilator.mk: refresh - didn't work tools/cdk/ocpidds/ocpidds/src/dataTypeTest.cxx: warning tools/local/tester/bin/src/ocpi-rcc-test_main.cxx: remove slashes in transport URLs

+++ core/container/hdl_container/impl/src/HdlContainer.cxx
@@ -1198,7 +1198,7 @@ namespace OCPI {
 	      base &= ~(pagesize - 1);
 	      top &= ~(pagesize - 1);
 	      fprintf(stderr, "DMA Memory is NOT page aligned.  Now %llu at 0x%llx\n",
-		      top - base, base);
+		      (unsigned long long)(top - base), (unsigned long long)base);
 	    }
           }
           done = true;
